siphon off allocation code + tests
